Remove default plugins

+++ src/init-j2digital-wp.sh
@@ -2,9 +2,17 @@
 
 set -e
 
-FOREGROUND=0 ./init-wp.sh
+DISABLE_FOREGROUND=1 ./init-wp.sh
 
-ln -s /wp-theme/dist /wp/wp-content/themes/j2digital
+if [[ ! -e /wp/wp-content/themes/j2digital ]]; then
+  ln -s /wp-theme/dist /wp/wp-content/themes/j2digital
+fi
+
+chown -R wp:wp /wp/wp-content/themes/j2digital
+
+rm -rf /wp/wp-content/themes/twenty*
+rm -rf /wp/wp-content/plugins/akismet
+rm -rf /wp/wp-content/plugins/hello.php
 
 tail -f /dev/null
